requirements.txt
fastapi
uvicorn[standard]
openai
anthropic
python-dotenv
requests
supabase
psycopg2-binary